fix restore size when float is used
it was technically possible for geometries to not be integer, but
a string to int parse was done, so teoretically string.toInt() can
fail and reset position/size of the plasmoid.
It may be the cause of some of the positions lost bugs.
This was particularly evident and triggerable in plasma mobile.
on palsma mobile side was also fixed by making suire the grid size is
always int, however technically the bug is here too as well

+++ components/containmentlayoutmanager/gridlayoutmanager.h
@@ -27,11 +27,11 @@ class AppletsLayout;
 class ItemContainer;
 
 struct Geom {
-    int x;
-    int y;
-    int width;
-    int height;
-    int rotation;
+    qreal x;
+    qreal y;
+    qreal width;
+    qreal height;
+    qreal rotation;
 };
 
 class GridLayoutManager : public AbstractLayoutManager
